Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I'm not sure the advantage of taking this, as it also requires us to maintain it. In particular, these custom names are really marketing oriented and could change on a whim, making maintenance a problem.

In addition, I suspect that there several languages/locales for each of these that should be created, further burdening maintenance.

So it seems to me using the overlay capabilities of the platform is more flexible and preferred. Of course, I'm willing to listen to counter arguments.

-- Wink

Patch-set: 1
Reviewer: Wink Saville <1001401@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0