DISPLAY TEXT clr msg after delay, expects result code OK
Send RES_ID_TIMEOUT response if StkDialogActivity is sent to the
background due to another acitivty has replaced it as forground
activity. When a callback call interferes and closes the
StkDialogActivity, the session still needs to be ended. If the
session is not ended in a controlled way, STK will end up in a
wrong state. If STK is in wrong state it will be unable to
process subsequent callbacks.

Change-Id: Ib6519ce15225b1c7be83251c6461fb14a4680200