No response sent when backlight is OFF
When the LCD backlight is OFF while receiving a
PLAY TONE command, the onDestroy() method is called
immediately since the app is in the background.
In onDestroy() the STOP message is removed from the
handler queue which leads to no response is sent back to
StkAppService. This causes the RIL layer to send
TR 'ME currently unable to process command' back to the SIM.
Solution is to check if a response is sent before
removing any messages from the handler queue.

Change-Id: I7ac2e99b85b38576acfe72370aa55850cdc304a0