Stk: Add support for Select Item - Default Item
As per 3GPP 11.14, if the Select Item is received
with Item Identifier set to one of the items, then
the ME should select/highlight the default item.
Currently, default item is not selected/highlighted.

This patch fixes the issue by highlighting the default
item by setting the background color.

Change-Id: Ieda19d9dfb65c796450a9108f19bd139f516099b
Author: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 61229