Stk: These changes are related to DISPLAY Text pro cmd
The following issues related to DISPLAY_TEXT are fixed in this change:

1. Send the Terminal Response if ResponseNeeded qualifier is set, after
the default duration else display message for ever.

2. Currently the STK App is sending "screen busy" terminal response for
Subsequent display text proactive commands if a previous display text is
being displayed. Changes are made to update the dialog box with the
subsequent display text and not to send the screen busy.
(Refer to ETSI TS 102.384 section 27.22.4.1.4.4.2) .

3. When a high priority Display Text proactive command with infinite
time out is sent,user has to clear this message. The Terminal Response for
this command will be sent from Stk framework itself. Stk app receives
this message and displays it until user clears it. In between other
proactive commands like Play Tone can be processed parallely. The
Terminal Response for Play Tone command is not sent if the Display Text message
is not cleared because of improper handling. This is corrected.

4. When a high priority, wait for user clear and no timeout Display Text
Proactive command is received UE should display the text until user
clears it. While the text is being displayed if a next Proactive command is
received, it should be acted upon immediately.

5.Don't consider screen busy if STK related activity is displayed.

We could receive display text proactive command while an STK
activity is already being displayed. In this case, don't check for
idle screen before displaying the activity.

Bug: 10490429
Change-Id: I8b6438359026a93237a0fff9f13d36c91bfb9f47