Update patch set 1
Patch Set 1:

> Here's a description of the bug for which this code was added:
 > 
 > After performing a call through STK menu, phone is not refreshing
 > menu options and screen.
 > 
 > Steps to Reproduce:
 > 1) Turn on phone with a SIM inserted
 > 2) Go to STK
 > 3) Select "Mis Servicios" option from menu (STK Menu 1)
 > 4) Select any option which performs a phone call like "Atn.
 > Clientes" (STK Menu 2)
 > 5) Accept performing the call (STK Menu 3)
 > 6) Verify call is connected and the hang up
 > 7) Phone should go back to previous menu or main menu (STK Menu 5)
 > 8) Select another option to perform another call (STK Menu 6)
 > 9) Verify call is connected
 > 
 > Expected behavior:
 > After step 7, Phone should go to main menu or stay in previous
 > menu. Phone should be able to perform any of the options selected.
 > 
 > Actual behavior:
 > After step 7, phone is showing menu "STK Menu 2" but actually is in
 > main menu. Phone is not reflecting/showing the right menu after
 > phone call is terminated. You can verify this by choosing any of
 > the options available and they won't match with current menu shown.

Previously the launch mode of StkMenuActivity was "singleTop", so only one instance could exist and behaved both as the main menu or secondary menu. launchMenuActivity() was called NOT just to launch the activity BUT also to refresh/recreate the activity in order to have it with the appropriate state. Even if it is called many times, the number of activity's instances was always 1.

However, with move to the multiple instance architecture, if we call launchMenuActivity() multiple times, then multiple 'MAIN' menu activities can co-exist in the activity stack. That is the issue we faced and want to fix. Say that now you are seeing 'MAIN' menu and push the back key, but 'MAIN' menu is still here because two instances were created because of the multiple calls to launchMenuActivity(). That is the current issue.

The bug you describe should not happen. StkMenuActivity instantiated as the secondary menu is finished in onStop() when the in-call screen comes to the foreground.

Patch-set: 1
Label: Verified=0