Update patch set 1
Patch Set 1:

Can you elaborate a little more on the scenario? Is this really about displaying dialog if command comes while sleeping?

Your change is in onStop(). Is this to avoid calling finish()? How is the responseNeeded check helping here?

Regarding, "dialog shall not be treated as
pending dialog if the immediate response option is specified", I am not sure what pending dialog is but this seems to contradict the earlier statement that dialog should be displayed.

Patch-set: 1