LAUNCH_BROWSER cmd should not be handled by Setup Wizard application
WebDialogActivity implemented in com.google.android.setupwizard.util is
Launched instead of a browser if Intent.ACTION_VIEW is broadcasted
with http/https schemes when the phone setup has not been
completed. WebDialogActivity cannot always handle user operations
on a carrier's SIM activation page which was displayed
because of the request from SIM, so it's better NOT to broadcast the
intent if the phone setup has not completed yet.

Change-Id: I0ce7a4f87f477f13bd28ffb96bac23e75b67859d