Update patch set 1
Patch Set 1:

> Can you elaborate a little more on the scenario? Is this really
 > about displaying dialog if command comes while sleeping?
 > 
 > Your change is in onStop(). Is this to avoid calling finish()? How
 > is the responseNeeded check helping here?
 > 
 > Regarding, "dialog shall not be treated as
 > pending dialog if the immediate response option is specified", I am
 > not sure what pending dialog is but this seems to contradict the
 > earlier statement that dialog should be displayed.

Steps to reproduce the issue:
1. Let the phone enter sleep state
2. Receive a DISPLAY TEXT command with the immediate response option
3. Remove the screen lock and check whether or not the message dialog is displayed

Since the immediate response option is set at (2.), terminal response is sent back without user confirmation right after starting StkDialogActivity (see StkAppService.launchTextDialog).

Now, StkAppService.handleSessionEnd and consequently StkAppService.cleanUpInstanceStackBySllot destroy "pending activity" when the current session ends, but that means that StkDialogActivity will be destroyed in the above scenario, since the state of activity changes from onCreate() -> onStart() -> onResume() then immediately to onPause() -> onStop() when the lock screen exists on top of the screen - because onStop will call setPendingDialogInstance(this). To avoid this, we check "responseNeeded" which is set to false if immediate response option is used.

Patch-set: 1