The main menu shall not be launched when the current session ends
The main menu is unexpectedly launched sometimes when user attempts to
remove the STK application from the recent apps screen.

That can happen if user sees the secondary menu before moving to the
recent apps screen, for example. After user attempts to remove the STK
application from the recent apps, mMainActivityInstance is changed to
null when the main menu is destroyed first, then the end of the session
for the secondary menu is notified next. Because of that, the main menu
is automatically launched in launchMenuActivity().

Bug: 31325698
Change-Id: I2ec896e556bd307e58ccc21f9fe9592ef91b5f26