Support preset STK names and icons for MSIM model
This change provides an easy way to replace the STK application name and
icon displayed on the launcher activity with preset resources. There are
many SIM cards in the market that send SET-UP MENU command with no alpha
identifier. This change improves the usability of multi-SIM models in
such cases.

Bug: 31295893

Change-Id: I437fa761cdd8bf47eee660c7bb2828b84c91e8af