The STK menu activity cannot treat subsequent user input
The boolean field 'mAcceptUsersInput' is mistakenly left in some
use-cases after it is changed to false. That causes an issue that user
cannot do anything on the STK main menu.

The boolean field is changed to false when user clicks one of the menu
items on the STK main menu. It is changed back to true in onResume()
when screen transition happens. However, we cannot say that screen
transition always happens after selecting the item on the main menu.

That is why we have to introduce one more trigger to reset the value of
'mAcceptUsersInput' when the current session is ended.

Bug: 31288411
Change-Id: I15ac01269a474e1dbc04ad250762066dfdb70f1e