Update patch set 1
Patch Set 1:

This is just an additional explanation for this change.
The main part of the change is to do the following 3 things when the current session ends.

    mAcceptUsersInput = true;
    mProgressView.setIndeterminate(false);
    mProgressView.setVisibility(View.GONE);

The local broadcast is convenient for the interface between the service and the activity in this kind of cases.
It is ok to unregister the newly intoroduced receiver in onPause()
because onResume() will be executed before the next user input and onResume() will do all we want.

Patch-set: 1
Label: Verified=0