Timeout should be handled as expected even in the sleep mode
The timeout handler does not work as expected if the terminal goes into
the sleep mode. StkDialogActivity does not have the same issue because
the scheduled alarm is used instead of the handler with the delayed
message. The same solution should be applied to the other activities
still using the delayed message solution.

Change-Id: I9521c27410852760c608f2dd42589463dcc447b8