Update patch set 5
Patch Set 5: Code-Review+1

Let me merge the following change to this commit because the purpose of these changes are completely same. It looks that the following one was submitted internally before, but it is still not available on the master branch.

  Prevent NPE in StkDialogActivity due to missing StkAppService
  https://android-review.googlesource.com/201971

Please let me know if these changes should be separated.

Patch-set: 5
Reviewer: Yoshiaki Naka <1115438@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1