Incorrect slot ID can cause NullPointerException after power cycle
If user does power-cycle the phone when the STK menu activity is listed
on the overview screen, the STK menu activity appears again on the
overview screen after the phone is restarted. However, the instance
state is not saved in this case, so correct slot ID cannot be retrieved
when user attempts to go to the STK menu activity from the overview
screen. Unintentional behavior must be avoided.

Bug:
Test: manual - confirmed that NPE is not thrown if we apply this.

Change-Id: Ia8624c45b34abd6f8ea73424b03ce453d92711df