Close the dialog activity before new one is displayed
It is recommended to close the dialog displayed for DISPLAY TEXT command
with an immediate response object before new one is displayed in
accordance with the requirement described in GSMA PDATA.12 - SIM Toolkit
Device Requirements to improve Mobile Connect Customer Experience to
avoid confusion.

ETSI 102.223 also says that the terminal shall continue to display the
text until it receives a subsequent proactive command containing display
data in this scenario, so there is no reason to keep displaying the
former one.

Bug:
Test: Confirmed the expected behaviors in the relevant scenarios.

Change-Id: Ia87cb3f410f6c95dcc820450c3792b920556bb56