Update patch set 2
Patch Set 2:

The issue 1 is covered by the following change. Please check the code differences to determine which is better solution.

    The activity should not be left when the dialog disappears
    https://android-review.googlesource.com/428380 (Bug: 63270020)

The following change fixes the issue 2. Sony recommends you choose this because it is not good to display a small icon on the action bar of the dialog. Please check both changes. Thanks.

    Terminal shall display icon specified by DISPLAY TEXT command
    https://android-review.googlesource.com/441133 (Bug: 63914016)

Patch-set: 2
CC: Yoshiaki Naka <1115438@85c56323-6fa9-3386-8a01-6480fb634889>