Improve the logic to determine the idle mode
The definition of the idle state is that idle screen is available in the
foreground of the screen according to ETSI TS 102.384 TC 27.22.4.22 "SET
UP IDLE MODE TEXT" and TC 27.22.7.6 "Idle screen available event". It is
required to detect which application is in the foreground to pass the
test cases. Intent.ACTION_SCREEN_OFF is not appropriate for that.

Bug:
Test: Confirmed that the test cases mentioned above can be passed.

Change-Id: I479100761843e91a7f1269d5258aef1684791c57