Update patch set 2
Patch Set 2:

Dear Grace,

There is no such guidelines as far as I checked the standard documents and the conformance test specification,
so I believe that both contributions can pass the conformance test cases.
However, the icon can be self-explanatory with no text message, so it should not be too small.
The following change has been merged for the otehr activity and it is similar to 'another contribution'.

  Icon must be placed near the text string on the input screen
  https://android-review.googlesource.com/451016

Best regards,
Yoshiaki

Patch-set: 2
CC: Yoshiaki Naka <1163305@85c56323-6fa9-3386-8a01-6480fb634889>