Update patch set 4
Patch Set 4:

We decided to remove the help option from "More options" because it is available as the context menu that you can see if you do long press any list item. The help option needs an index number of the list item, so the context menu is appropriate.

Patch-set: 4
CC: Yoshiaki Naka <1163305@85c56323-6fa9-3386-8a01-6480fb634889>