Update patch set 5
Patch Set 5:

Dear Jordan,

Thanks for your precious time for this review.

We added the drawables to reserve the resource IDs for icons which will be added as overlay resources.
Can we add overlay icons by using <add-resource type="drawable"> and assign appropriate resource IDs to them?

We will try it and update the solution if it is available, but please give us a time.
It will not be finished in a few days.

Best regards,
Yoshiaki

Patch-set: 5