Fix WindowLeaked issue happening in StkInputActivity
WindowManager says that StkInputActivity can have a leaked window while
user is changing the multi-window mode. That is because finish() is
called in onResume() during the mode change. The basic design of this
activity is that it should not be finished but stay in the background if
the dialog activity is launched on it. It should be finished when the
current session ends or it is completely covered by another activity
except for the dialog activity.

Change-Id: I1374583ef8ae3521d77bc05090505f8a64d1c0ee