Update patch set 1
Patch Set 1:

So far we have found not issue with the following fix candidate.
Please review this, too ?
I think that it is not the best solution but better than this change - Change 401132.

    Fix incorrect handling of pending activities
    https://android-review.googlesource.com/519215

Change 401132 is just about to change the pending activity mechanism to a kind of FIFO queue with depth 2.
It is still possible to push the activity out from the queue without finish().
One example is;

    Stpp 1.) Open the main menu.
    Step 2.) Select an item and receive SELECT ITEM command from SIM card.
    Step 3.) Move to the home screen, then wait for the timeout of SELECT ITEM command.
    Step 4.) TERMINAL RESPONSE is sent back to SIM card due to the timeout, then receive GET INPUT command next.
    Step 5.) Push the back key while seeing StkInputActivity to end the current session.

After this procedure, we will see StkMenuActivity (secondary) created at the step 2 even if we apply Change 401132.
That is because StkInputActivity calls setPendingActivityInstance(this) multiple times.
The same two instances of StkInputActivity pushes StkMenuActivity out of the queue,
so StkMenuActivity loses a chance to be finished by StkAppService in this use-case.

Now I think that it might be better to get rid of the "pending activity" mechanism and will investigate it next,
but Change 519215 might be an acceptable solution for now.
Please let me know if you have any question.

Patch-set: 1