Fix terminal response not sent during sim hotswap.
In a case, where single sim is removed and reinserted,
terminal response for the subsequent display text/select
item proactive command is not sent, which results in GCF
test case failure.

Fix: When sim card is absent initialize the CatService instance
to null, so that StkAppService always use the latest CatService
instance when sim card becomes present.

Bug: 74049285
Change-Id: Ie2c48ce9d392339638604b760e022def8098d186