Fix to dismiss ToneDialog activity on tone timeout.
For shorter tone timeout duration in milliseconds,
ToneDialog activity creation takes more than the tone
duration itself, so the ToneDialog activity will not
get destroyed.

This is because StkAppService is maintaining the play
tone timer and ToneDialog activity is not yet created
when StkAppService tries to finish the ToneDialog activity,
resulting in PlayTone GCF test case 27.22.4.5.1.1 failure.

Fix: Move the code to finish the ToneDialog activity from
     StkAppService.java to ToneDialog.java.
   - ToneDialog.java will keep track of the playtone timeout
     duration and finishes off the activity on timeout.

Bug: 35615464
Change-Id: I42a20ba898f7b15c1b42af458d86c4bc0c753350