Update patch set 4
Patch Set 4:

Dear Jordan,

I found another strange behavior introduced by Change: 172563 on the partner git.
This is just a note to tell the issue to you, so I will not address it this time here.

    Step 1.) Open the main menu of the SIM Toolkit application from the application launcher.
    Step 2.) Enter the following command from PC several times.

        $adb shell am start -n com.android.stk/.StkMain

The result is that StkMenuActivity is instantiated multiple times in restoreInstanceFromStackBySlot().
That means that many instances of StkMenuActivity are created in the activity stack,
so I had to push the back key many times to close all the instances of the main menu.

It should not be a critical issue, but one of the misteries in Change: 172563 on the partner git.

Best regards,
Yoshiaki

Patch-set: 4
Reviewer: Yoshiaki Naka <1163305@85c56323-6fa9-3386-8a01-6480fb634889>