Update patch set 2
Patch Set 2: Code-Review+1

Dear Reviewers,

The current design of the SIM Toolkit application expects that the outside of the dialogs is transparent,
so let me recommend to merge this next.

I retested this change on the top of the latest merged commit below today.
It works as expected and no other issue was found in the test.

    056ae69 Merge "The main menu shall not be launched when the current session ends

Best regards,
Yoshiaki

Patch-set: 2
Reviewer: Yoshiaki Naka <1163305@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1