Update patch set 1
Patch Set 1: Code-Review-1

Dear Reviewers,

We confirmed that there are some issues in this solution.

First, the instance state is correctly saved and restored while recreating the instance,
so the delayed message will be lost.

Another issue is that ToneDialog sometimes remains for a long time even after we apply this fix
probably because of a delay of the broadcast message.
We can upload another solution which utilizes AlarmManager as with StkDialogActivity.

Best regards,
Yoshiaki

Patch-set: 1
Reviewer: Yoshiaki Naka <1163305@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1