Finish pending activity or dialog when the current session ends
If a pending activity or a pending dialog exists, it must be finished
when the current session ends regardless of the current command.
Otherwise, the pending one may remain on the screen without accepting
any user operation. cleanUpInstanceStackBySlot() immediately exits
without doing anything if the current command is set to null in the
current implementation, but it must be just to avoid null ponter access.

Bug:
Test: Manually confirmed that pending activity is finished as expected.

Change-Id: I90c12d1a0694bf8c35f68e9429136935392f9496