Override the label of the SIM Toolkit application dynamically
The list of the menu items provided by SET UP MENU command shall be a
part of the menu system of the terminal and the alpha identifier shall
be used for the title of the menu in accordance with the clause 6.4.8
SET UP MENU in ETSI 102.223. The terminal shall be able to display the
alpha identifier as the lable of the launcher activity of SIM Toolkit.
This feature should be disabled if more than one main menu is available
at the same time.

Change-Id: I215de452f14aaa04b7389197aac03c1d50257ae0