The card state ABSENT can be expected in REFRESH with UICC Reset
When REFRESH command with UICC Reset comes from the SIM card, CatService
is disposed and the broadcast intent AppInterface.CAT_ICC_STATUS_CHANGE
is sent with the card state 'ABSENT', then the menu information of SIM
Toolkit is purged and the launcher activity is removed in StkAppService.
We do not need to do the same when the broadcast is sent with the card
state 'PRESENT'.

Change-Id: Ifd815ea19e42cf7dd4d58c9c3292b6c8662864a9