The dialog should be canceled if user goes to the home screen
SIM Toolkit can suddenly launch message dialog in accordance with
request from SIM card. User might not understand that the dialog is
associated with SIM Toolkit application, so many users will be unable to
come back to the dialog once they push the home button and go to the
home screen. GSMA PDATA.12 recommends that the terminal returns TERMINAL
RESPONSE with the cause "Proactive UICC session terminated by the user"
if user goes to the home screen while the dialog is displayed.

Change-Id: I64254a0ea555510655803bb830848516cdf86406