The current session shall be terminated if user goes to the idle screen
User might see a dialog, an input screen or an item selection list
screen which is displayed in accordance with request from the SIM card.
It is possible that the user does not understand that the dialog or
screens are associated with SIM Toolkit application, so many users will
probably be unable to come back to them once the user pushes the home
button and goes to the idle screen. GSMA PDATA.12 recommends that the
terminal returns TERMINAL RESPONSE with the cause "Proactive UICC
session terminated by the user" if user goes to the home screen.

Bug: 113367485
Test: Confirmed that the device can behave as mentioned in PDATA.12.

Change-Id: I64254a0ea555510655803bb830848516cdf86406