Avoid TransactionTooLargeException in onSaveInstanceState()
Menu object can have a bitmap resource 'titleIcon' and its size can be
big, but it is not allowed to save big data as a part of the instance
state. The instance of the menu class is retrieved always when the
activity is resumed, so saving it obviously makes no sense.

Bug: 118178431
Test: The exception is never thrown after applying this change.

Change-Id: I5b03ea614238c8e5c07d76626e3307120e1f2241