Differentiate home screen from others implemented in the same activity
Home screen and overview screen cannot be distinguished if those are
implemented in the same activity like Launcher in com.android.launcher3
as the current implementation determines that the foreground screen is
the home if it has the home category. Using ACTION_CLOSE_SYSTEM_DIALOGS
must be better solution for detecting the home screen.

Bug: 130142364
Test: Performed manual test cases which includes regression tests.

Change-Id: I2db9181aadbb59fd2ae620cdb41f3e3bcd132e6b