Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This level of test doesn't look sufficient - it's not exercising the feature and I don't the the failure will generate a build or boot error.  Is this a CP from a CL which got more testing?

It's a backport from ag/6163781, merged on 25 Jan, when the bug was marked fixed. See that bug for some context.

This CL just adds a permission, which is unlikely to break anything, and restores the status quo - the app will continue to be able to launch activities from the background.

If there are specific scenarios that you'd like to verify that would require someone with more familiarity with the details of this app.

Patch-set: 1