Remove the task of SIM Toolkit when it becomes no longer needed
The problem is that user can reinstantiate the activity which already
sent TERMINAL RESPONSE from the card if user find the task of SIM
Toolkit on the overview screen and select it. The unnecessary task must
be removed from the overview screen and some flags for launching
activity should be removed instead.

Change-Id: I74b078241b5954f138f509360e6942ca20064d43