Remove SIM Tookit from the launcher screen if no main menu is available
SIM Toolkit has a logic to remove itself from the launcher screen once
it becomes unnecessary, but the condition is wrong. It is removed when
all the SIM cards go into ABSENT state in the current implementation,
but it must be executed when the number of active main menu becomes 0.

Change-Id: I8091bf6ea82df873335cce3477a6c759735ed79b